ai/requirements.txt
psycopg2-binary==2.9.9
xgboost==2.1.3
scikit-learn==1.5.2
onnx==1.17.0
onnxruntime==1.20.1
onnxmltools==1.12.0
skl2onnx==1.17.0
anthropic==0.39.0
pandas==2.2.3
numpy==1.26.4
boto3==1.35.71
